[#17] Feat: DataManager 에서 데이터 로드하는 로직 분리
DataManager 역할 재구현 필요
